inewave/nwlistop/viol_vazmin.py
from inewave.nwlistop.modelos.blocos.usina import Usina
from inewave.nwlistop.modelos.arquivos.arquivousinapatamar import (
    ArquivoUsinaPatamar,
)
from inewave.nwlistop.modelos.viol_vazmin import ViolVazminAnos


class ViolVazmin(ArquivoUsinaPatamar):
    """
    Armazena os dados das saídas referentes à violação de defluência
    mínima da usina.

    Esta classe lida com as informações de saída fornecidas pelo
    NWLISTOP e reproduzidas nos `viol_vazmin00x.out`, onde x varia conforme
    a usina em questão.

    """

    BLOCKS = [
        Usina,
        ViolVazminAnos,
    ]
